fix: trainopts.ts tests
